Mark extents for unallocated files and for deleted files (duh) and make sure the inums output in the disk map are converted to being signed, to allow for negative values.

+++ src/main.cpp
@@ -113,7 +113,7 @@ void outputDiskMap(const std::string& diskMapFile, std::shared_ptr<LbtTskAuto> w
           if (!firstFile) {
             file << ", ";
           }
-          file << "{\"inum\":" << f.first << ", \"id\":" << f.second << "}";
+          file << "{\"inum\":" << static_cast<int64_t>(f.first) << ", \"id\":" << f.second << "}";
           firstFile = false;
         }
         file << "]}";
